requirements.txt
ncoreparser
schedule
requests
python-dotenv
